refactor: Participant의 Player 코드 분리
