+++ src/main/java/domain/participants/Player.java
@@ -1,8 +1,10 @@
 package domain.participants;
 
 import domain.BettingMoney;
+import domain.Command;
 
 public class Player extends Participant {
+    private static final int BLACK_JACK = 21;
 
     private final BettingMoney bettingMoney;
 
@@ -13,6 +15,18 @@ public class Player extends Participant {
 
     public int getMoney() {
         return bettingMoney.getBettingMoney();
+    }
+
+    public boolean isOverPlayerBlackJack() {
+        return getCardsSum() >= BLACK_JACK;
+    }
+
+    public boolean canDrawCard(Command command) {
+        return !isOverPlayerBlackJack() && isCommandYes(command);
+    }
+
+    public boolean isCommandYes(Command command) {
+        return command.equals(Command.YES);
     }
 
 }
