+++ src/test/java/domain/DealerTest.java
@@ -24,7 +24,7 @@ public class DealerTest {
 
         Card card = new Card(CardNumber.ACE, CardPattern.SPADE);
         dealer.addCard(card);
-        assertThat(dealer.getPlayerCards().get(0)).isEqualTo(card);
+        assertThat(dealer.getCards().get(0)).isEqualTo(card);
     }
 
     @Test
